refactor (Action) : 파라미터 변경
