+++ src/main/java/subway/menu/Action.java
@@ -1,31 +1,35 @@
-package subway;
+package subway.menu;
 
+import subway.controller.ControllerFactory;
 import subway.controller.SubwayController;
 
 public enum Action {
 
     REGISTER("등록") {
         @Override
-        public void action(SubwayController controller) {
+        public void action(Menu menu) {
+            SubwayController controller = ControllerFactory.of(menu);
             controller.save();
         }
     },
     DELETE("삭제") {
         @Override
-        public void action(SubwayController controller) {
+        public void action(Menu menu) {
+            SubwayController controller = ControllerFactory.of(menu);
             controller.delete();
         }
     },
     SEARCH("조회") {
         @Override
-        public void action(SubwayController controller) {
+        public void action(Menu menu) {
+            SubwayController controller = ControllerFactory.of(menu);
             controller.findAll();
         }
     },
     BACK("") {
         @Override
-        public void action(SubwayController controller) {
-            // 아무것도 안함
+        public void action(Menu menu) {
+            //아무것도 안함
         }
     };
 
@@ -35,7 +39,7 @@ public enum Action {
         this.action = action;
     }
 
-    abstract public void action(SubwayController controller);
+    abstract public void action(Menu menu);
 
     @Override
     public String toString() {
