+++ src/main/java/subway/menu/LineMenu.java
@@ -32,7 +32,7 @@ public enum LineMenu implements Menu {
     public Menu run() {
 
         try {
-            action.action(controller);
+            action.action(this);
             return MainMenu.QUIT;
         } catch (Exception e) {
             OutputView.printErrorMessage(e);
