+++ src/test/java/subway/Test.java
@@ -1,0 +1,2 @@
+package subway;public class Test {
+}
